pyproject.toml
[project]
name = "cuddle-market-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.9",
    "django>=5.2.5",
    "django-cors-headers>=4.7.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "drf-spectacular>=0.28.0",
    "ruff>=0.12.8",
]

[tool.ruff.lint.per-file-ignores]
"config/settings/__init__.py" = ["F403", "F405"]
"config/settings/dev.py" = ["F403", "F405"]
"config/settings/prod.py" = ["F403", "F405"]